Move API members to files grouped by REST resources
